Refactor ManifestReader and auto update manifest
Moved ManifestReader from DataReader.py into its own module and renamed class as Manifest.
Added new method for updating manifest.csv in place from a given folder
of raw data csv files using filename convention that includes
unique_data_id in the filename. Also, moved create_list method from
LoadData.py into Manifest.py.

Added code to get_data_api that automatically updates manifest.csv for
each unique_data_id that is updated from that script.

The changes were confirmed through local update of the database and via
test_load_data.py. The latter required updating the test data files used
to reflect recent updates.
