+++ python/housinginsights/ingestion/DataReader.py
@@ -122,75 +122,6 @@ class HIReader(object):
         return None
 
 
-class ManifestReader(HIReader):
-    """
-    Adds extra functions specific to manifest.csv. This is the class that
-    should be used to read the manifest and return it row-by-row.
-    """
-
-    _include_flags_positive = ['use']
-    _include_flags_negative = ['pending', 'exclude', 'superseded']
-
-    def __init__(self, path='manifest.csv'):
-        super().__init__(path)
-        self.unique_ids = {}  # from the unique_id column in the manifest
-
-        # validate the manifest
-        if not self.has_unique_ids():
-            raise ValueError('Manifest has duplicate unique_data_id!')
-
-    def __iter__(self):
-        self._length = 0
-        self._counter = Counter()
-        with open(self.path, 'r', newline='') as data:
-            reader = DictReader(data)
-            self._keys = reader.fieldnames
-            for row in reader:
-                self._length += 1
-
-                #parse the date into proper format for sql
-                try:
-                    _date = dateparser.parse(row['data_date'],dayfirst=False,
-                                             yearfirst=False)
-                    row['data_date'] = datetime.strftime(_date, '%Y-%m-%d')
-                except ValueError:
-                    row['data_date'] = 'Null'
-
-                #return the row
-                yield row
-
-    def has_unique_ids(self):
-        """
-        Verifies that every value in the manifest column 'unique_data_id' is
-        in fact unique. This makes sure that we can rely on this column to
-        uniquely identify a source CSV file, and and can connect a record in
-        the SQL manifest to the manifest.csv
-
-        :return: true if passes validation; false otherwise
-        """
-        self.unique_ids = {}
-
-        for row in self:
-            if row['unique_data_id'] in self.unique_ids:
-                return False
-            else:
-                #don't add flags that won't make it into the SQL database
-                if row['include_flag'] in ManifestReader._include_flags_positive:
-                    self.unique_ids[row['unique_data_id']] = 'found'
-        return True
-
-    def get_manifest_row(self, unique_data_id):
-        """
-        Returns the row for the given unique_data_id else 'None' if not in
-        manifest.
-        """
-        for row in self:
-            use = row['include_flag'] == 'use'
-            if row['unique_data_id'] == unique_data_id and use:
-                return row
-        return None
-
-
 class DataReader(HIReader):
     """
     Reads a specific data file. This file must be associated with a specific
