+++ python/housinginsights/ingestion/__init__.py
@@ -1,6 +1,6 @@
 
 
-from .DataReader import ManifestReader, DataReader
+from .DataReader import DataReader
 #from .functions import load_meta_data, check_or_create_sql_manifest
 
 #Replace this method?
@@ -9,7 +9,7 @@ from .SQLWriter import HISql, TableWritingError
 
 
 #TODO add to this
-__all__ = [	'ManifestReader',
+__all__ = [	'Manifest',
 			'DataReader',
 			'load_meta_data',
 			'check_or_create_sql_manifest',
