--- python/tests/test_data/temp_manifest.csv
@@ -1,26 +1,0 @@
-include_flag,destination_table,unique_data_id,update_method,data_date,encoding,local_folder,s3_folder,filepath,notes
-____City Development Data__,,,,Null,,,,,
-superseded,building_permits,building_permits_2016_old,manual,2017-01-01,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/building_permits/2016/Building_Permits_2016.csv,
-skip,building_permits,building_permits_2014,manual,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/building_permits_2014.csv,
-skip,building_permits,building_permits_2015,manual,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/building_permits_2015.csv,using files from opendata batch download
-skip,building_permits,building_permits_2016,manual,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/building_permits_2016.csv,using files from opendata batch download
-skip,building_permits,building_permits_2017,manual,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/building_permits_2017.csv,using files from opendata batch download
-_____Preservation Catalog_____,,,,Null,,,,,
-use,project,project,manual,2017-03-15,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/preservation_catalog/20170315/Project.csv,
-use,subsidy,subsidy,manual,2017-03-15,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/preservation_catalog/20170315/Subsidy.csv,
-____Crime Data____,,,,Null,,,,,
-skip,crime,crime_2011,manual,2011-12-31,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/crime/mpd-crime-map/20170403/2011-all.csv,skipping for now until we are able to display older data
-skip,crime,crime_2012,manual,2012-12-31,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/crime/mpd-crime-map/20170403/2012-all.csv,skipping for now until we are able to display older data
-skip,crime,crime_2013,manual,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/crime_2013.csv,skipping for now until we are able to display older data
-skip,crime,crime_2014,manual,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/crime_2014.csv,skipping for now until we are able to display older data
-use,crime,crime_2015,api,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/crime_2015.csv,
-use,crime,crime_2016,api,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/crime_2016.csv,
-use,crime,crime_2017,api,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/crime_2017.csv,
-____WMATA Data____,,,,Null,,,,,
-superseded,wmata_dist,wmata_dist_20170215,api,2017-02-15,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/wmata/20170215/dist.csv,
-superseded,wmata_info,wmata_info_20170215,api,2017-02-15,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/wmata/20170215/wmatainfo.csv,
-use,wmata_dist,wmata_dist_20170215,api,2017-05-24,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/wmata/20170215/dist_updated_headers.csv,
-use,wmata_info,wmata_info_20170215,api,2017-05-24,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170524/wmata_stops.csv,
-____DATA FOR APPENDING____,,,,Null,,,,,
-use,project,dchousing_project,api,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/dchousing_project.csv,append to 20170315/project.csv
-use,subsidy,dchousing_subsidy,api,06/06/2017,latin-1,../../../data,https://s3.amazonaws.com/housinginsights/,raw/apis/20170606/dchousing_subsidy.csv,append to 20170315/Subsidy.csv
